refactor: getDiscountTypeChecks to utils files
